Refactor mocking sublime and sublime_plugin
