+++ plugins/ctags.py
@@ -8,7 +8,6 @@ import mmap
 import os
 import re
 import subprocess
-import sys
 
 from subprocess import check_output
