+++ plugins/tests/mock_sublime_plugin.py
@@ -1,0 +1,19 @@
+"""
+Mock module for ``sublime_plugin`` in Sublime Text.
+"""
+
+all_callbacks = {
+    'on_load': []
+}
+
+
+class WindowCommand(object):
+    pass
+
+
+class TextCommand(object):
+    pass
+
+
+class EventListener(object):
+    pass
